Added Lexical Categories, as promised
The Lexical Categories system is exactly as previously described. It's awesome
and hopefully it works because it isn't tested yet. I completely removed every
reference to openSymbols and now Word and Terminal are implemented as
LexicalCategory sub(case)class/object. Plus, I factored the tokenization out of
the Grammar class and into its own. Parsable now makes the choice of how to
tokenize a string, as makes the most sense. Tokens are specified by Parsables,
but are only useful (as far as I can see, for now) on terminals.