Finally finished coding solution algorithm
The first iteration of the code is done, but it's not working yet. I set up a
very basic test case ("John kissed Gary") but it's throwing an exception because
some ID is not found. This was kind of expected, and it's good that it's
throwing the exception instead of just silently failing: I SHOULD be able to
guarantee that I never try to access an ID that doesn't exist. The next step is
to actually guarantee that..