Lots of parsing goodness
In the LFG package, I factored out all of the Parser objects to a Parsables
object that you can choose to import. I also added parsing commands namespaced
within the parsing package that should be (from now on) the preferred method of
doing parsing. Basically, now I'm being faithful to the way I implemented the
Parsable thing in the first place: It's a type class! So I will migrate all of
my parsing objects to be implicit and separate from the classes (and partner
objects) they're parsing for. This is nice because now they can have parser
objects too (with important, non-parsing functionality) and there doesn't have
to be any fighting over names.

Also, I made lexical specifications parsable and it works AWESOMELY! So now test
code reads like it was written down by a linguist! kind of. you know what I
mean. It's great. That's the point.