Nicened up parsing test code
Now there is a clear relationship between whitespace and the intended braces in
the LFG lexical category specification. This makes room for a
whitespace-sensitive preprocessor.

Also... the parse time of that long string is appreciable. Quite
unfortunate...hahaha. I can only imagine what it'd be like to parse something
like that if it was specified in LFG... speaking of which, that's another thing
on the todo list. First, create a nice system for creating symbolic objects that
are parsable with LFG. Then, try specifying and parsing some stuff!