Fixed TODO.txt and added parsing for no-tokens
I had forgotten to add the ability to actually return the parse trees for the
empty string if that was part of your grammar and you were just parsing an empty
list of tokens. Now that's done, and the todo list is updated: I've done parsing
with the empty category, and as a side effect mitigated the problem with a
cyclic grammar.