Added lots of toStrings, and parsable inv-app
These toStrings should probably be moved into parser objects: it's nice to have
parsers that can go both ways (string -> obj, obj -> string).

We get a stack overflow on "John kissed" so it looks like there may be some sort
of bug with the implementation. ...Or, empty categories are just causing the
execution to nest itself ridiculously much in the parsing process. The next step
to solving this problem will probably involve a bunch of logging to try to
figure out what's going wrong.