Fixed test code
And all of the tests pass! However, a strange problem arose with the case object
CNFEmpty. It was failing to initialize for some reason. Changing it to a
parameterless case class worked, but is slightly uglier than the object
implementation because it only needs to be a singleton. So much for generalized
algebraic data types.