Add agenda-based parsing
Also updated to a different implementation of sorted streams.
Actually this seems to have introduced a subtle bug in the old CKY
parser, which causes the LFG test suite to fail. I've narrowed it down
to IdentifyingExpression[RelativeIdentifier] not working.
So don't use this version for the old CKY parser.

.. I'm going to change it out soon anyway ..